k8s-bot/requirements.txt
python-telegram-bot==21.0
kubernetes==30.1.0
python-dotenv==1.0.1
streamlit==1.40.0
pandas>=2.2.3
numpy>=2.1.0
httpx>=0.25.0,<0.28.0